Pin ruby to version 1.9.3 to make heroku consistent with my dev-test environment.
LocalSupport rake test tasks use ruby 1.9.3, but heroku used 1.9.2

This follows the recommendation from the Heroku installer:
-----> WARNINGS:
       you have not declared a Ruby version in your Gemfile.
       To set your Ruby version add this line to your Gemfile:"
       ruby '1.9.2'"
       See https://devcenter.heroku.com/articles/ruby-versions for more information.

At this moment, rake tasks used on my development VM are

    /usr/bin/ruby1.9.1 -S rspec ....
    /usr/bin/ruby1.9.1 -S bundle exec cucumber  --profile default

Although thanks to the mysteries of Debian

    $ /usr/bin/ruby1.9.1 -v
    ruby 1.9.3p194 (2012-04-20 revision 35410) [i486-linux]

Note that this pin is currently ignored in the dev-test environment, per this
message: "Ignoring `ruby` directive. This is a feature added to Bundler 1.2.0
and higher. Please upgrade if you would like to use it. "