implement client side url validation
Implement client side validation through activerecord validation. Added
a method to add url protocol and removed dependency on url_validation
gem in favour of client_side_validation gem. I believe this is a good
way of simplifying the code and making it more transparent, because
server-side and client-side validation are defined in the same place.