Remove unecessary markup
Removed the divs surrounding the label tags as they're unecessary since we don't
need the form-group class. These could auto generate in the future if we ever
decide to use the rails-bootstrap-forms gem.
https://github.com/bootstrap-ruby/rails-bootstrap-forms#usage

Removing the divs and class does not change the structure as it is now.
The form-group class only purpose is to add bottom-margin and form control.
https://v4-alpha.getbootstrap.com/components/forms/#form-groups <- I'm confident
this applies to version 3 as well.