150247799 layout of volunteer op form is crooked (#516)
* Remove bootstrap override

The input tag is inline by default. Removed the inline-block declaration
for the input tag that was overriding the bootstrap style.

This is necessary to allow bootstrap classes to control the UX and to have
a responsive web page

* Add bootstrap class to form tags

Remove invalid and overriding bootstrap classes to replace with bootstrap's
own classes for web forms.

The field class was invalid and not defined and wide_form class added
unnecessary top margin causing display irregularities. Removed the custom
classes in exchange for bootstrap classes.

[Finishes #150247799]

* Remove unecessary markup

Removed the divs surrounding the label tags as they're unecessary since we don't
need the form-group class. These could auto generate in the future if we ever
decide to use the rails-bootstrap-forms gem.
https://github.com/bootstrap-ruby/rails-bootstrap-forms#usage

Removing the divs and class does not change the structure as it is now.
The form-group class only purpose is to add bottom-margin and form control.
https://v4-alpha.getbootstrap.com/components/forms/#form-groups <- I'm confident
this applies to version 3 as well.