Update install documentation (#533)
* Add issue during npm install

During npm install bower flags for an invalid-meta data issue, but aborts
install process. Added temporary solution to resolve this issue.

Until we can migrate away from bower we needed a solution to work past this
issue for a smoother onboarding process.

* Update ruby version

Ruby version was updated and this needs to be reflected in the
installation notes.

* Update development environment setup instructions

Added link to gorails to setup a development environment.

gorails has a very good documentation that is kept updated for setting
up a development environment accross many different platforms.