Add call to tweet new volunteer opportunities
Update the run method to tweet new volunteer opportunities.

Perhaps this should be in a seperate rake tasks to tweet new volunteer
opportunities, but since two external API's are being called and the
timing of such actions can be delayed for various reasons we instead add
a call to our TwitterApi to tweet all new volunteer opportunities from
DoIt as our last command in this task which ensures the calls to the
external API's (DoIt & Twitter) are consecutive and that we don't miss
tweeting new volops due to timing issues that are out of our control.