Assign coordinates to nil when invalid
Change the assignment statement for setting the latitude and longitude
to nil instead of 0.0

0.0 is falsey in javascript and custom marker's are not created, however
the map will still move to center on latitude and longitude coordinates
of 0.0 even though there is not a custom marker, because the coordinates
for centering the map are retrieved from the database. Setting the value
to nil (empty string will work too) solves the issue of the map jumpting
to invalid coordinates.

[Finishes #150890866]