Create test to catch invalid lat/lng assignments
Add tests that check imported volunteer opportunities are being assigned
valid map coordinates.

Volunteer opportunities that are imported with nil coordinates or the
GeoCoder is unable to assign valid coordinate are then assigned a
coordinate of 0.0, 0.0 and the marker is placed in the middle of the
world. These tests ensure that all imported volunteer opportunities are
persisted to the database with valid coordinates.