Update tests for TwitterApi
Removed expectation that a parsed tweet will be 280 characters. This can
cause an unexpected failure if/when Twitter changes the allowed text
limit.

Parse method was changed to a private method and should not qualify for
unit testing as it is used by the tweet method. Add tests to check for a
sent tweet on both a long and short message.