Don't persist volops outside of Harrow (#755)
* Update fixtures location name to Harrow

Location name is a unique identifer allowing us to filter volops that
are not in Harrow. This is needed so that we don't persist volops
outside of Harrow.

* Add fixture volop outside of Harrow

Added new volop that is outside of Harrow.

So that we can check we're not persisting data outside of Harrow.

* Don't persist volops outside of Harrow

Checking location_name before persisting and skip when it is not Harrow.

* Add constants for central harrow location

* Replace magic numbers with Location constants

* Update feature spec

Updated feature spec as it is only matching 3 volops since we do a check
on location_name as well.

Updated VCR file to have some Harrow location names.

* Update counts

Volops now check for valid location name

* Update vcr file

Location name needs to be valid in order to persist volop.

* Remove default nil coords

Removing all specs, cassettes, and default coords.

Having null coords for doit volops should not be an issue anymore now
that location_name is being used.

* Replace magic numbers with constants

* Refactor coordinate assignment

Refactoed coordinate assignment into seperate method.

This improves readability by stating what should be done by moving the
decision making to a seperate method. This also fixes the code climate
issue of an instruction line being to long.

* Explicitly state private method

* Refactor method to location class

Create method instantiates a location object that sets
latitude/longitude coordinates to either the passed in coords or default
location coords.

This was changed to increase readability and to keep methods that belong
together into one class.

* Setup Location class to get coordinates

* Set coordinates unless nil

Reachskills may send us nil coordinates and if they do then they will be
set to the default location upon object creation.

* Refactor persist doit vol op method

Refactored the persist volop method to seperate responsbility.

Made this change to minimize method responsibility

* Add missing op parameter

* Return model after populating attributes

* Refactor process doit method

Refactored main process method.

Hopefully this improves readability and is moving toward single
responsibility. Could probably use some more work.

* Refactor the next page request

To improve readability and single reponsibility

* Rename method

* Fix syntax