139801559 fix vol op posting to doit api (#802)
* Test for a Y/M/D date format

Added tests to check for a y/m/d date format.

Ensures random strings are not persisted in date fields as the doit API
will return an error if this occurs.

* Add format validation to the date fields

* Add tests to validate doit vol op date range

Added validation tests to check the advertise start and end dates meet
iso8601 standards.

The date needs to be a specific format when submitting a volunteer
opportunity to the doit api.

* Validate advertise dates

Added custom validations to check the start date is not in the past and
to check end date is not before the start date. Also moved format
validation to the custom method so that it can short circuit before the
comparison to ensure both operands are date type.

* Refactor validation methods

Extracted and removed the duplicate RegEx matchers.
Removed the extra step for going to the new vol op create page.

* Update test to use date class methods

Changed hardcoded future date to class method for grabbing current date
and tomorrows date.

This is changed so that the test doesn't fail after the year 9999 :-)

* Fix date comparison validation

Updated to compare Date objects instead of Time objects.

Comparing Time.current with a date string fails on edge cases when the
date is the same, but times are different or non-existent. For this test
we're not concerned with the time, but rather the date.

* Add placeholder text to display date format

* Fix date picker

* Fix test that is failing validation

Updated advertise start date to the current date.

Validation checks does not allow for start dates to be in the past.

* Refactor

Single level abstraction

* Change to single quoted strings

* Turn Rubocop's frown upside down.and sideways()